gemini midi

+++ src/main/java/org/chuck/core/ChuckDuration.java
@@ -5,17 +5,17 @@ import java.util.concurrent.TimeUnit;
 /**
  * Represents a duration in ChucK (e.g., 10::ms).
  */
-public record ChuckDuration(long samples) implements Comparable<ChuckDuration> {
-    public static ChuckDuration of(long samples) {
+public record ChuckDuration(double samples) implements Comparable<ChuckDuration> {
+    public static ChuckDuration of(double samples) {
         return new ChuckDuration(samples);
     }
 
     public static ChuckDuration fromMs(double ms, int sampleRate) {
-        return new ChuckDuration((long) (ms * sampleRate / 1000.0));
+        return new ChuckDuration(ms * sampleRate / 1000.0);
     }
 
     public static ChuckDuration fromSeconds(double seconds, int sampleRate) {
-        return new ChuckDuration((long) (seconds * sampleRate));
+        return new ChuckDuration(seconds * sampleRate);
     }
 
     public ChuckDuration plus(ChuckDuration other) {
@@ -24,6 +24,6 @@ public record ChuckDuration(long samples) implements Comparable<ChuckDuration> {
 
     @Override
     public int compareTo(ChuckDuration o) {
-        return Long.compare(this.samples, o.samples);
+        return Double.compare(this.samples, o.samples);
     }
 }
